PR090619

+++ Applications/DataExplorer/VtkVis/VtkCompositeFilter.cpp
@@ -5,7 +5,7 @@
  * \brief  Implementation of the VtkCompositeFilter class.
  *
  * \copyright
- * Copyright (c) 2012-2018, OpenGeoSys Community (http://www.opengeosys.org)
+ * Copyright (c) 2012-2019, OpenGeoSys Community (http://www.opengeosys.org)
  *            Distributed under a Modified BSD License.
  *              See accompanying file LICENSE.txt or
  *              http://www.opengeosys.org/project/license
@@ -33,8 +33,7 @@ VtkCompositeFilter::~VtkCompositeFilter()
     _outputAlgorithm->Delete();
 }
 
-
-float VtkCompositeFilter::GetInitialRadius() const
+double VtkCompositeFilter::GetInitialRadius() const
 {
     double bounding_box[6];
     static_cast<vtkPolyData*>(this->_inputAlgorithm->GetOutputDataObject(0))->GetBounds(bounding_box);
